% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signed.mean.R
\name{signed.mean}
\alias{signed.mean}
\title{take average of positive or negative values.}
\usage{
signed.mean(vector, type)
}
\arguments{
\item{vector}{input vector to be averaged.}

\item{type}{string; either "pos" or "neg".}
}
\value{
n.mean averaged positive or negative values from the vector.
}
\description{
this function averages positive or negative values over the entire length of the input vector.
}
\examples{
signed.mean(c(0, -1, 2, 5), type = "pos")
}

